switch to fibonacci as a library
