+++ web/controllers/page_controller.ex
@@ -15,19 +15,7 @@ defmodule Fawkes.PageController do
 
   def fibonacci(conn, _params) do
     {n,_} = Integer.parse(_params["nth"])
-    r = fastfib(n)
+    r = Fibonacci.fastfib(n)
     render conn, "fibonacci", [nnumber: n,result: r]
   end
-
-  defp fastfib(0), do: 0
-  defp fastfib(1), do: 1
-  defp fastfib(n) do
-    sqrt5 = :math.sqrt(5)
-    res = (1/sqrt5) * ( :math.pow((1+sqrt5)/2 , n) - :math.pow((1-sqrt5)/2, n) )
-    round(res)
-  end
-
-  defp fib(0), do: 0
-  defp fib(1), do: 1
-  defp fib(n), do: fib(n-1)+fib(n-2)
 end
